Fix problems where a peptide has only one non-missing value, by giving an error
- Data is median normalized, and then NaNs are replaced with zeros
- This means that after normalization, if there is only one non-missing
value for a peptide, the values for that peptide will be all zeros, at
least in the form input to the factor analysis.
- In the factor analysis, there is a correlation matrix calculated. If
one of the input peptides is always zero, one of the diagonal entries in
the correlation matrix will be zero, rather than one.
- This means the correlation matrix is not of full rank. The code does
not seem to be set up for this case.
- Returning an error seems reasonable.